decouple modalcontent from modal popup into its own component
