+++ frontend/src/modules/controllers/index.ts
@@ -20,4 +20,5 @@ export { default as ValueController } from './ValueClicker';
 export { default as AlarmModal } from './AlarmModal';
 export { default as ValueClicker } from './ValueClicker';
 export { default as ModalPopup } from './ModalPopup';
+export { default as ModalContent } from './ModalContent';
 export { default as SimpleTable } from './SimpleTable';
